fix customuserdetails

+++ src/main/java/com/mindhub/todolist/controllers/user_controllers/TaskController.java
@@ -24,16 +24,22 @@ public class TaskController {
     private CustomUserDetailsService customUserDetailsService;
 
     @GetMapping("/{id}")
-    public ResponseEntity<?> getTaskById(@PathVariable Long id) throws TaskExceptions, UserExceptions {
+    public ResponseEntity<?> getTaskById(@PathVariable Long id) throws TaskExceptions {
+        Long user_id = customUserDetailsService.getAuthUserId();
         Task task = taskServices.getTaskById(id);
-        customUserDetailsService.isTheUserAuth(task.getUser().getId());
+        if (!task.getUser().getId().equals(user_id)){
+            throw new TaskExceptions("Invalid User");
+        }
         TaskDTO taskDTO = taskServices.getTaskDTOById(id);
         return new ResponseEntity<>(taskDTO, HttpStatus.OK);
     }
 
     @PostMapping
-    public ResponseEntity<?> createTask(@RequestBody NewTask newTask) throws UserExceptions {
-        customUserDetailsService.isTheUserAuth(newTask.userModel().getId());
+    public ResponseEntity<?> createTask(@RequestBody NewTask newTask) throws TaskExceptions {
+        Long user_id = customUserDetailsService.getAuthUserId();
+        if (!newTask.userModel().getId().equals(user_id)){
+            throw new TaskExceptions("Invalid User");
+        }
         if(newTask.title().isBlank() || newTask.description().isBlank() || newTask.userModel().getId().toString().isBlank()) {
             return new ResponseEntity<>("Invalid Data", HttpStatus.BAD_REQUEST);
         }
@@ -43,8 +49,11 @@ public class TaskController {
 
     @PutMapping("/{id}")
     public ResponseEntity<?> updateTaskById(@RequestBody UpdateTask updateTask, @PathVariable Long id) throws TaskExceptions, UserExceptions {
+        Long user_id = customUserDetailsService.getAuthUserId();
         Task task = taskServices.getTaskById(id);
-        customUserDetailsService.isTheUserAuth(task.getUser().getId());
+        if (!task.getUser().getId().equals(user_id)){
+            throw new TaskExceptions("Invalid User");
+        }
         if (updateTask.title().isBlank()) {
             return new ResponseEntity<>("El titulo no puede estar en blanco", HttpStatus.BAD_REQUEST);
         }
@@ -53,14 +62,13 @@ public class TaskController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<?> deleteTaskById(@PathVariable Long id) throws TaskExceptions, UserExceptions {
-        try {
+    public ResponseEntity<?> deleteTaskById(@PathVariable Long id) throws TaskExceptions {
+            Long user_id = customUserDetailsService.getAuthUserId();
             Task task = taskServices.getTaskById(id);
-            customUserDetailsService.isTheUserAuth(task.getUser().getId());
+            if (!task.getUser().getId().equals(user_id)){
+                throw new TaskExceptions("Invalid User");
+            }
             taskServices.deleteTaskById(id);
             return new ResponseEntity<>(HttpStatus.NO_CONTENT);
-        } catch (TaskExceptions | UserExceptions exceptions) {
-            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
-        }
     }
 }
